Add TAG_Long_Array (id 0x0C) support
Minecraft 1.12 introduced TAG_Long_Array; vanilla world saves now use
it for biome storage and packed block-state palettes. Until now
TrueCraft.Nbt rejected the tag-id at the NbtBinaryReader range check
(NBT tag type out of range: 12).

- New TrueCraft.Nbt/Tags/NbtLongArray.cs mirroring NbtIntArray: id 0x0C,
  long[] payload, big-endian per-element write, deep-clone-on-construct,
  Skip/Read/Write paths that reject negative lengths.
- NbtTagType gains LongArray = 0x0C.
- NbtBinaryReader.ReadTagType bumps its upper bound to 0x0C.
- NbtWriter gains BeginList/BeginList-named upper-bound bumps and four
  WriteLongArray overloads matching the WriteIntArray surface.
- NbtReader gains TAG_Long_Array cases in HasLength, ReadValue path,
  Skip path, ReadAsTag path, and the boxed-value path. ToString skips
  the value for LongArray like it does for IntArray/ByteArray.
- NbtCompound.ReadTag and NbtCompound.SkipTag dispatch to NbtLongArray.
- NbtList.ListType range check accepts LongArray; NbtList.ReadTag and
  NbtList.SkipTag dispatch to NbtLongArray.
- NbtTag.GetCanonicalTagName returns "TAG_Long_Array".
- NbtTag gains LongArrayValue { get; } convenience accessor mirroring
  IntArrayValue.
- Serialization/NbtSerializer maps long[] <-> NbtTagType.LongArray.

Tests: new Test.TrueCraft.Nbt/LongArrayTests.cs covers enum value,
construction (default + cloning), indexer, deep-copy semantics, the
uncompressed and GZip+ZLib round-trip, big-endian on-the-wire layout
sanity, negative-length rejection, and a list-of-arrays-inside-compound
nested case. Test count: 95 -> 105.

+++ TrueCraft.Nbt/NbtBinaryReader.cs
@@ -34,7 +34,7 @@ namespace TrueCraft.Nbt
             int type = ReadByte();
             if (type < 0) throw new EndOfStreamException();
 
-            if (type > (int) NbtTagType.IntArray) throw new NbtFormatException("NBT tag type out of range: " + type);
+            if (type > (int) NbtTagType.LongArray) throw new NbtFormatException("NBT tag type out of range: " + type);
 
             return (NbtTagType) type;
         }
